Fix wrong int cast.

+++ codecCtx.go
@@ -285,7 +285,7 @@ func (this *CodecCtx) Channels() int {
 }
 
 func (this *CodecCtx) SetBitRate(val int) *CodecCtx {
-	this.avCodecCtx.bit_rate = C.int64_t(val)
+	this.avCodecCtx.bit_rate = C.int(val)
 	return this
 }
